+++ test/resampler/point.js
@@ -22,7 +22,8 @@ describe('Upsample a 16bit 8kHz file', function() {
     fs.readFileSync(path + "16-bit-8kHz-noBext-mono.wav"));
 
   // Convert to another sample rate
-  wav.toSampleRate(16000, {method: 'point', clip: 'periodic', LPForder: 2});
+  wav.toSampleRate(16000,
+    {method: 'point', clip: 'periodic', LPForder: 2, LPF: true});
 
   // Write the file
   fs.writeFileSync(